ResumeParsing/requirements.txt
gunicorn
flask
flask_cors
flasgger
json
os
gc
time
uuid
dotenv
google-genai
pdf2image
PIL